backend/requirements.txt
aenum==3.1.15
fastapi==0.103.1
mypy==1.5.1
Pillow==10.0.1
pycodestyle==2.11.0
pydantic==1.10.0a1
pylint==2.17.5
python-dotenv==1.0.0
python-multipart==0.0.6
scipy==1.11.1
sphinx-rtd-theme==1.2.2
sphinxcontrib-napoleon==0.7
uvicorn==0.23.2